get drugbank annotation from unichem ftp
One of the inclusion criteria to find a drug in Open Targets is that the same drug exists in DrugBank.
This check was done using a ChEMBL ID/DrugBank ID table from 5 years ago whose origin I didn't know.

Because of this issue https://github.com/opentargets/issues/issues/4167 , I was having a look at longer term solutions that would allow us to have up to date mappings. 

ChEMBL has crossreferences with DrugBank on their pages because they use the Unichem API to get them. I've just ran into this old issue of ours explaining that the obsolete look up tables we have are coming from Unichem FTP. So now the source for the PIS rule is not the latest file on the bucket, but the FTP directory itself (not versioned as far as i can see, always pointing to latest).

The format of the latest file is **the same** and we have over a **33% increase** in the number of cross references (8986 vs 6523).